requirements.txt
altgraph==0.17.2
bottle==0.12.19
bottle-websocket==0.2.9
Eel==0.14.0
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.1
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyparsing==2.4.7
pypng==0.0.21
PyQRCode==1.2.1
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
